+++ sdk/application/Content/ContentApplicationValidator.js
@@ -17,6 +17,8 @@ const ContentApplicationModel = require("./ContentApplicationModel");
  * @property {number} [pageNo] - The page number to navigate through the given
  *   set of results. Default value is 1.
  * @property {number} [pageSize] - The number of items to retrieve in each page.
+ * @property {string} [tags] - Blogs retrieve based on the list of tags passed.
+ * @property {string} [search] - Blogs retrieve based on the title.
  */
 
 /**
@@ -129,6 +131,8 @@ class ContentApplicationValidator {
     return Joi.object({
       pageNo: Joi.number(),
       pageSize: Joi.number(),
+      tags: Joi.string().allow(""),
+      search: Joi.string().allow(""),
     });
   }